fix: aligned error messages with replit

+++ src/main.rs
@@ -8,21 +8,29 @@ async fn get(storage: Storage, path: web::Path<(String,)>) -> Result<impl Respon
     if let Some(value) = value {
         Ok(HttpResponse::Ok().body(value))
     } else {
-        Err(error::ErrorNotFound("Not found."))
+        Err(error::ErrorNotFound(""))
     }
 }
 
 #[delete("/{key}")]
 async fn del(storage: Storage, path: web::Path<(String,)>) -> Result<impl Responder, Error> {
-    storage.delete(&path.0).await?;
-    Ok(HttpResponse::Ok())
+    if storage.contains_key(&path.0).await? {
+        storage.delete(&path.0).await?;
+        Ok(HttpResponse::NoContent())
+    } else {
+        Err(error::ErrorNotFound(""))
+    }
 }
 
 #[post("/")]
 async fn add(storage: Storage, req_body: String) -> Result<impl Responder, Error> {
+    if req_body.is_empty() {
+        return Err(error::ErrorBadRequest(""));
+    }
+
     let mut body_set = req_body.split('=');
-    let key = body_set.next().ok_or_else(|| error::ErrorBadRequest("Key Required."))?;
-    let value = body_set.next().ok_or_else(|| error::ErrorBadRequest("Key Required."))?;
+    let key = body_set.next().ok_or_else(|| error::ErrorBadRequest(""))?;
+    let value = body_set.next().ok_or_else(|| error::ErrorBadRequest(""))?;
     storage.set_bytes(key, value).await?;
     Ok(HttpResponse::Ok())
 }
